refactor(dir): split reno ui from room ui
